+++ tests/doc_index/base_classes/test_configs.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Type
 from pydantic import Field
 
 from docarray import BaseDocument
-from docarray.storage.abstract_doc_store import BaseDocumentIndex
+from docarray.storage.abstract_doc_index import BaseDocumentIndex
 from docarray.typing import NdArray
